fix: decision making
