+++ packages/valory/services/impact_evaluator_local/service.yaml
@@ -124,7 +124,6 @@ models:
       transaction_service_url: ${TRANSACTION_SERVICE_URL:str:https://safe-transaction-mainnet.safe.global/api/v1/messages/{message_hash}/}
       wveolas_address: ${WVEOLAS_ADDRESS:str:0x4039B809E0C0Ad04F6Fc880193366b251dDf4B40}
       tweet_consensus_wveolas: ${TWEET_CONSENSUS_WVEOLAS:int:2000000}
-      sync_on_ceramic_data: ${SYNC_ON_CERAMIC_DATA:bool:false}
 ---
 public_id: valory/ledger:0.19.0
 type: connection
