+++ packages/valory/skills/ceramic_read_abci/skill.yaml
@@ -141,7 +141,6 @@ models:
       slash_threshold_amount: 10000000000000000
       light_slash_unit_amount: 5000000000000000
       serious_slash_unit_amount: 8000000000000000
-      sync_on_ceramic_data: true
     class_name: Params
   requests:
     args: {}
