+++ packages/valory/skills/decision_making_abci/tasks/read_stream_preparation.py
@@ -40,13 +40,14 @@ class ReadCentaursPreparation(TaskPreparation):
         yield
         updates = {
             "read_stream_id": self.params.centaurs_stream_id,
-            "read_target_property": "centaurs_data",
+            "sync_on_ceramic_data": False,
         }
         return updates, self.task_event
 
     def _post_task(self):
         """Preparations after running the task"""
         updates = {
+            "centaurs_data": self.state.ceramic_data,
             "current_centaur_index": 0,  # reset the centaur index after reading
         }
         yield
@@ -69,13 +70,15 @@ class ReadContributeDBPreparation(TaskPreparation):
         yield
         updates = {
             "read_stream_id": self.params.ceramic_db_stream_id,
-            "read_target_property": "ceramic_db",
+            "sync_on_ceramic_data": False,
         }
         return updates, self.task_event
 
     def _post_task(self):
         """Preparations after running the task"""
         updates = {}
+        # Load the stream data into the db model
+        self.ceramic_db.load(self.state.ceramic_data)
         yield
         return updates, None
 
@@ -96,12 +99,12 @@ class ReadManualPointsPreparation(TaskPreparation):
         yield
         updates = {
             "read_stream_id": self.params.manual_points_stream_id,
-            "read_target_property": "score_data",
+            "sync_on_ceramic_data": False,
         }
         return updates, self.task_event
 
     def _post_task(self):
         """Preparations after running the task"""
-        updates = {}
+        updates = {"score_data": self.state.ceramic_data}
         yield
         return updates, None
