+++ packages/valory/skills/decision_making_abci/tasks/task_preparations.py
@@ -30,12 +30,14 @@ class TaskPreparation:
     task_name = None
     task_event = None
 
-    def __init__(self, synchronized_data, params, logger, now_utc, behaviour) -> None:
+    def __init__(self, synchronized_data, params, context, now_utc, behaviour) -> None:
         """Init"""
         self.name = ""
         self.synchronized_data = synchronized_data
         self.params = params
-        self.logger = logger
+        self.logger = context.logger
+        self.state = context.state
+        self.ceramic_db = context.ceramic_db
         self.now_utc = now_utc
         self.behaviour = behaviour
         self.set_config()
