+++ packages/valory/skills/impact_evaluator_abci/skill.yaml
@@ -197,7 +197,6 @@ models:
       transaction_service_url: https://safe-transaction-mainnet.safe.global/api/v1/messages/{message_hash}/
       wveolas_address: '0x4039B809E0C0Ad04F6Fc880193366b251dDf4B40'
       tweet_consensus_wveolas: 2000000
-      sync_on_ceramic_data: false
     class_name: Params
   randomness_api:
     args:
